Drop android source sets from binary-util
