+++ gradle/libs.versions.toml
@@ -1,4 +1,5 @@
 [versions]
+androidx-startup = "1.1.1"
 androidx-test-core = "1.5.0"
 androidx-test-runner = "1.5.2"
 
@@ -18,6 +19,7 @@ kotlinx-cli = "0.3.6"
 okio = "3.5.0"
 
 [libraries]
+androidx-startup-runtime = { module = "androidx.startup:startup-runtime", version.ref = "androidx-startup" }
 androidx-test-core = { module = "androidx.test:core", version.ref = "androidx-test-core" }
 androidx-test-runner = { module = "androidx.test:runner", version.ref = "androidx-test-runner" }
