--- library/binary-util/api/android/binary-util.api
@@ -1,9 +1,0 @@
-public abstract interface annotation class io/matthewnelson/kmp/tor/binary/util/InternalKmpTorBinaryApi : java/lang/annotation/Annotation {
-}
-
-public final class io/matthewnelson/kmp/tor/binary/util/OSInfo$Companion {
-}
-
-public final class io/matthewnelson/kmp/tor/binary/util/ProcessExt {
-}
-
