+++ library/binary-util/api/binary-util.api
@@ -1,0 +1,12 @@
+public final class io/matthewnelson/kmp/tor/binary/util/AndroidSdkIntKt {
+}
+
+public abstract interface annotation class io/matthewnelson/kmp/tor/binary/util/InternalKmpTorBinaryApi : java/lang/annotation/Annotation {
+}
+
+public final class io/matthewnelson/kmp/tor/binary/util/OSInfo$Companion {
+}
+
+public final class io/matthewnelson/kmp/tor/binary/util/ProcessExt {
+}
+
