--- library/binary-util/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/binary/util/internal/-JvmPlatform.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2023 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.kmp.tor.binary.util.internal
-
-import io.matthewnelson.kmp.tor.binary.util.InternalKmpTorBinaryApi
-
-@JvmSynthetic
-@InternalKmpTorBinaryApi
-@Suppress("NewApi", "NOTHING_TO_INLINE")
-internal actual inline fun Process.forciblyDestroyInternal(): Process = destroyForcibly()
