+++ library/binary/src/androidUnitTest/kotlin/io/matthewnelson/kmp/tor/binary/OSInfoUnitTest.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.kmp.tor.binary
+
+import io.matthewnelson.kmp.tor.binary.util.InternalKmpTorBinaryApi
+import io.matthewnelson.kmp.tor.binary.util.OSHost
+import io.matthewnelson.kmp.tor.binary.util.OSInfo
+import kotlin.test.Test
+import kotlin.test.assertFalse
+
+@OptIn(InternalKmpTorBinaryApi::class)
+class OSInfoUnitTest {
+
+    @Test
+    fun givenOSInfo_whenAndroidUnitTestNonEmulator_thenHostIsNotLinuxAndroid() {
+        // Should register as the host machine running unit tests, not android
+        // This will fail if test being executed via Termux, but willing to take
+        // that chance..........
+        assertFalse(OSInfo.INSTANCE.osHost is OSHost.Linux.Android)
+    }
+}
