+++ tools/check-publication/build.gradle.kts
@@ -62,6 +62,7 @@ kmpConfiguration {
             sourceSetMain {
                 dependencies {
                     implementation("$group:binary:$version")
+                    implementation("$group:binary-util:$version")
                 }
             }
         }
